python/requirements.txt
ccxt
redis
python-dotenv